+++ NaCl/lib/NaCl/Box.hs
@@ -60,7 +60,7 @@ import qualified NaCl.Box.Internal as I
 --
 --     2. /Random/. You generate a random nonce every time you encrypt a message.
 --     Since the nonce is large enough, the chances of you using the same
---     nonce twice are negligible. For useful helpers, see @Crypto.Random@,
+--     nonce twice are negligible. For useful helpers, see @Crypto.Sodium.Random@,
 --     in <https://hackage.haskell.org/package/crypto-sodium crypto-sodium>.
 --
 --     In either case, you need to be able to provide the same nonce when decrypting,